Refactor tests
Split into smaller files and functions
More strategic tests
