+++ test/Jamfile.v2
@@ -41,10 +41,16 @@ run test_convert.cpp ;
 run test_env.cpp ;
 run test_env.cpp : : : <define>BOOST_NOWIDE_TEST_INCLUDE_WINDOWS=1 : test_env_win ;
 run test_fs.cpp : : : <library>/boost/filesystem//boost_filesystem/<warnings-as-errors>off ;
+run test_filebuf.cpp  ;
+run test_filebuf.cpp : : : <define>BOOST_NOWIDE_USE_FILEBUF_REPLACEMENT=1 <target-os>windows:<build>no : test_internal_filebuf ;
+run test_ifstream.cpp ;
+run test_ofstream.cpp ;
 run test_fstream.cpp ;
 run test_fstream.cpp : : : <define>BOOST_NOWIDE_USE_FILEBUF_REPLACEMENT=1 <target-os>windows:<build>no : test_internal_fstream ;
 run test_fstream_cxx11.cpp  ;
 run test_fstream_cxx11.cpp : : : <define>BOOST_NOWIDE_USE_FILEBUF_REPLACEMENT=1 <target-os>windows:<build>no : test_internal_fstream_cxx11 ;
+run test_fstream_special.cpp  ;
+run test_fstream_special.cpp : : : <define>BOOST_NOWIDE_USE_FILEBUF_REPLACEMENT=1 <target-os>windows:<build>no : test_internal_fstream_special ;
 run test_iostream.cpp ;
 run test_stackstring.cpp ;
 run test_stat.cpp ;
